backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-multipart==0.0.9
Pillow==10.3.0
mediapipe==0.10.14
numpy==1.26.4
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
